feat: add websocket server to http-server file
