+++ package.json
@@ -3,25 +3,25 @@
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js"
+      "start": "node ./index.ts",
+      "start:dev": "nodemon ./src/index.ts"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.22"
+      "@types/node": "^20.11.19",
+      "@types/uuid": "^9.0.8",
+      "@types/ws": "^8.5.10",
+      "nodemon": "^2.0.22",
+      "ts-node": "^10.9.2",
+      "ts-node-dev": "^2.0.0",
+      "typescript": "^5.3.3",
+      "uuid": "^9.0.1"
     },
     "dependencies": {
-        "@types/node": "^20.11.20",
-        "@types/uuid": "^9.0.8",
-        "@types/ws": "^8.5.10",
-        "ts-node": "^10.9.2",
-        "ts-node-dev": "^2.0.0",
-        "typescript": "^5.3.3",
-        "uuid": "^9.0.1",
-        "ws": "^8.8.0"
+      "ws": "^8.16.0"
     }
-}
+  }
+
